ml-engine/requirements.txt
pandas
pymongo
implicit
scipy
fastapi
uvicorn
joblib
python-dotenv
pytz
